// frontend/types.ts
export enum UserRole {
  ADMIN = "admin",
  EMPLOYEE = "employee",
}

export enum UserStatus {
  ACTIVE = "active",
  INACTIVE = "inactive",
}

export interface User {
  id: string;
  name: string;
  email: string;
  mobile: string;
  role: UserRole;
  status: UserStatus;
  createdAt: string;
}

export enum ApplicationStatus {
  NOT_STARTED = "not_started",
  DRAFT = "draft",
  SUBMITTED = "submitted",
  APPROVED = "approved",
  REJECTED = "rejected",
}

export interface Application {
  employeeId: string;
  fullName: string;
  email: string;
  mobile: string;
  gender: string;
  highestQualification: string;
  collegeName: string;
  yearOfPassing: string;
  percentageOrCgpa?: string;
  technicalSkills: string[];
  otherSkills: string[];
  status: ApplicationStatus;
  submittedAt?: string;
  updatedAt: string;
  googleDriveLink?: string;
  submittedDocs?: string[];
}